pyproject.toml
[project]
name = "onion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "pillow>=12.1.1",
    "python-multipart>=0.0.18",
    "scikit-learn>=1.6.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "uvicorn>=0.34.0",
]